not using pandas panels anymore.
Original idea to use pandas panels was convenience, and be able to
perform certain vector operations -

For the kind of filtering we need, often we are better of 'iterating'
over the data (transpose becoming very expensive), so vector operations
didn't provide any benefit. So simply using the 'dict' of dataframes and
iterating over the keys.

Just wrote a simple ema function to see how long it takes. Just
experimental though for now.